@@ -1,46 +1,20 @@
-﻿using System;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Serilog;
-using Serilog.Events;
+using Microsoft.Extensions.Hosting;
 
 namespace DanClarkeBlog.Web
 {
-    public static class Program
+    public class Program
     {
         public static void Main(string[] args)
         {
-            try
-            {
-                var configuration = new ConfigurationBuilder()
-                    .AddEnvironmentVariables()
-                    .Build();
-
-                var builder = WebHost.CreateDefaultBuilder(args)
-                    .UseStartup<Startup>()
-                    .UseSerilog()
-                    .UseConfiguration(configuration);
-
-                InitLogging();
-
-                builder.Build().Run();
-            }
-            finally
-            {
-                Log.CloseAndFlush();
-            }
+            CreateHostBuilder(args).Build().Run();
         }
 
-        private static void InitLogging()
-        {
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Warning()
-                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
-                .Enrich.FromLogContext()
-                .WriteTo.LiterateConsole()
-                .WriteTo.ApplicationInsights(Environment.GetEnvironmentVariable("APPINSIGHTS_INSTRUMENTATIONKEY"), TelemetryConverter.Traces)
-                .CreateLogger();
-        }
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
